[3D Models] added wavefront object import capability
x) can load simple wavefron object 3D models from file
  currently only vertices, normals and texture coordinates can be loaded
x) added teapot 3D model

+++ src/ARiftControl.cpp
@@ -30,21 +30,21 @@ ARiftControl::ARiftControl()
     leftCameraParameters_.Nyc = 90.0f;
     leftCameraParameters_.z = -250.0f;
   }
-  leftCameraParameters_.p9 = 0.030943828391667e+04;
-  leftCameraParameters_.p8 = 0.282942249820711e+04;
-  leftCameraParameters_.p7 = 1.150239951453818e+04;
-  leftCameraParameters_.p6 = 2.755836619048961e+04;
-  leftCameraParameters_.p5 = 4.338342090721455e+04;
-  leftCameraParameters_.p4 = 4.700973643667896e+04;
-  leftCameraParameters_.p3 = 3.520272074605963e+04;
-  leftCameraParameters_.p2 = 1.746733371991807e+04;
-  leftCameraParameters_.p1 = 0.555549279942498e+04;
-  leftCameraParameters_.p0 = 0.132980078899012e+04;
-  leftCameraParameters_.c = 1.000374008613590;
-  leftCameraParameters_.d = -9.611375125204626e-05;
-  leftCameraParameters_.e = -1.127978103886716e-04;
-  leftCameraParameters_.xc = 2.119393300155207e+02;
-  leftCameraParameters_.yc = 3.546241749915297e+02;
+  leftCameraParameters_.p9 = (float)(0.030943828391667e+04);
+  leftCameraParameters_.p8 = (float)(0.282942249820711e+04);
+  leftCameraParameters_.p7 = (float)(1.150239951453818e+04);
+  leftCameraParameters_.p6 = (float)(2.755836619048961e+04);
+  leftCameraParameters_.p5 = (float)(4.338342090721455e+04);
+  leftCameraParameters_.p4 = (float)(4.700973643667896e+04);
+  leftCameraParameters_.p3 = (float)(3.520272074605963e+04);
+  leftCameraParameters_.p2 = (float)(1.746733371991807e+04);
+  leftCameraParameters_.p1 = (float)(0.555549279942498e+04);
+  leftCameraParameters_.p0 = (float)(0.132980078899012e+04);
+  leftCameraParameters_.c = (float)(1.000374008613590);
+  leftCameraParameters_.d = (float)(-9.611375125204626e-05);
+  leftCameraParameters_.e = (float)(-1.127978103886716e-04);
+  leftCameraParameters_.xc = (float)(2.119393300155207e+02);
+  leftCameraParameters_.yc = (float)(3.546241749915297e+02);
   leftCameraParameters_.height = 752.0f;
   leftCameraParameters_.width = 480.0f;
